Moved VerifySignature to the AuthorizationController
