--- source/marketparticipant/Energinet.DataHub.MarketParticipant.EntryPoint.WebApi/Controllers/AuthorizationVerificationController.cs
@@ -1,65 +1,0 @@
-﻿// Copyright 2020 Energinet DataHub A/S
-//
-// Licensed under the Apache License, Version 2.0 (the "License2");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using Energinet.DataHub.Core.App.Common.Abstractions.Users;
-using Energinet.DataHub.MarketParticipant.Application.Commands;
-using Energinet.DataHub.MarketParticipant.Application.Commands.Actors;
-using Energinet.DataHub.MarketParticipant.Application.Commands.Authorization;
-using Energinet.DataHub.MarketParticipant.Application.Commands.Delegations;
-using Energinet.DataHub.MarketParticipant.Application.Security;
-using Energinet.DataHub.MarketParticipant.Domain.Model;
-using Energinet.DataHub.MarketParticipant.Domain.Model.Delegations;
-using Energinet.DataHub.MarketParticipant.Domain.Model.Permissions;
-using Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Revision;
-using Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Security;
-using Energinet.DataHub.RevisionLog.Integration.WebApi;
-using MediatR;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-
-namespace Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Controllers;
-
-[ApiController]
-[Route("authorization")]
-public class AuthorizationVerificationController : ControllerBase
-{
-    private readonly IMediator _mediator;
-    private readonly IUserContext<FrontendUser> _userContext;
-
-    private readonly IAuthorizationService _authorizationService;
-
-    public AuthorizationVerificationController(IAuthorizationService authorizationService, IMediator mediator, IUserContext<FrontendUser> userContext)
-    {
-        _authorizationService = authorizationService;
-        _mediator = mediator;
-        _userContext = userContext;
-    }
-
-    [HttpPost("VerifySignature")]
-    [AllowAnonymous]
-    public async Task<ActionResult> VerifySignatureAsync()
-    {
-        //var command = new VerifySignatureCommand();
-
-        //await _mediator
-        //    .Send(command)
-        //    .ConfigureAwait(false);
-
-        return Ok();
-    }
-}
